backend/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
python-multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4
sqlalchemy==2.0.9
psycopg2-binary==2.9.6
pydantic==1.10.7
pillow==9.5.0
numpy==1.24.3
opencv-python==4.7.0.72
geojson==3.0.1
pyproj==3.5.0
pandas==2.0.1
shapely==2.0.1
rasterio==1.3.6
geopandas==0.13.0
pytorch==2.0.0
torchvision==0.15.1
ultralyticsplus==0.0.23
pyyaml==6.0
tqdm==4.65.0
